requirements.txt
numpy
pillow
openai
python-dotenv